Randomly Pin a 6-Pin Challenge Lock
This code was written for the March 2017 Lockpicking Advent Calendar.  The goal of the calendar is to promote Locksport, grow the community and... HAVE FUN!

The output of this code will randomize a pinning for a 6 chamber practice lock. Please check the notes next to the code to customize the code for your lock / pins