Added RandomLockAndRandomizeAPracticeLock.py
Code that chooses a random lock from a user-defined list to pick or randomly generates pinnings of a lock with user-defined chamber amounts, spring types, key pin types, key bittings and driver pin types.

Code was written for days 19-31 of the March 2017 Lockpicking Advent Calendar to randomize a challenge lock each day at the end of the month.  The goal of the Calendar is to promote locksport, grow the community and have fun.

All code is written by TheEJM3 and can be used, modified and distributed as long as proper credit is given to the original creator.